+++ wagtaildraftail/fields.py
@@ -11,7 +11,7 @@ class DraftailTextField(models.TextField):
     empty_values = list(EMPTY_SERIALIZED_JSON_VALUES)
 
     def __init__(self, *args, **kwargs):
-        self.editor = kwargs.pop('editor', 'default_draftjs')
+        self.editor = kwargs.pop('editor', 'default_draftail')
         super(DraftailTextField, self).__init__(*args, **kwargs)
 
     def formfield(self, **kwargs):