+++ wagtaildraftail/utils.py
@@ -1,9 +1,10 @@
 from __future__ import absolute_import, unicode_literals
 
 from django.conf import settings
+from django.template.defaultfilters import filesizeformat
 from django.utils. module_loading import import_string
-from draftjs_exporter.defaults import BLOCK_MAP
 
+from draftjs_exporter.defaults import BLOCK_MAP
 
 _exporter_config = None
 
@@ -29,3 +30,17 @@ def get_exporter_config():
         }
 
     return _exporter_config
+
+
+def get_document_meta(document):
+    """
+    :type document: wagtail.wagtaildocs.models.Document (or subclass)
+    :param document: the document
+    :rtype: dict
+    :return: the document size and extension
+    """
+
+    return {
+        'extension': document.file_extension.lower(),
+        'size': filesizeformat(document.file.size)
+    }